Split layer persistence parser
