+++ doc/reference/file-map.md
@@ -116,6 +116,8 @@
 - `src/document/persistence.c`
 - `src/document/persistence_json.c`
 - `src/document/persistence_json.h`
+- `src/document/persistence_layers.c`
+- `src/document/persistence_layers.h`
 - `src/document/persistence_write.c`
 - `src/app/workspace_service.c`
 - `src/app/application_file_actions.c`
